rename format_pfx to parse_pfx, adjust parse logic
we're really parsing here, not formatting: the network may be invalid
from sources provided upstream. Thus we return None for an invalid
network and handle accordingly.
