+++ kartograf/util.py
@@ -131,7 +131,7 @@ def wait_for_launch(wait):
         time.sleep(1)
 
 
-def format_pfx(pfx):
+def parse_pfx(pfx):
     """
     Attempt to format an IP network or address.
     If invalid, return None.
@@ -163,12 +163,13 @@ def get_root_network(pfx):
     Extract the top-level network from an IPv4 or IPv6 address.
     Returns the value as an integer.
     """
-    network = format_pfx(pfx)
-    v = ipaddress.ip_network(network).version
-    if v == 4:
-        return int(network.split(".", maxsplit=1)[0])
+    network = parse_pfx(pfx)
+    if network:
+        v = ipaddress.ip_network(network).version
+        if v == 4:
+            return int(network.split(".", maxsplit=1)[0])
 
-    root_net = network.split(":", maxsplit=1)[0]
-    if root_net:
-        return int(root_net, 16)
-    return 0
+        root_net = network.split(":", maxsplit=1)[0]
+        if root_net:
+            return int(root_net, 16)
+    return None
