+++ tests/bogon_test.py
@@ -1,5 +1,5 @@
 from kartograf.bogon import is_bogon_pfx, is_bogon_asn, extract_asn, is_out_of_encoding_range
-from kartograf.util import format_pfx
+from kartograf.util import parse_pfx
 
 def test_special_asns():
     special_cases = [0, 112, 23456, 65535, 4294967295]
@@ -77,18 +77,5 @@ def test_valid_ipv6_prefixes():
         "2a03:2880::/32"     # Facebook
     ]
     for prefix in valid_prefixes:
-        network = format_pfx(prefix)
+        network = parse_pfx(prefix)
         assert is_bogon_pfx(network) is False
-
-def test_invalid_prefixes():
-    invalid_prefixes = [
-        "not.a.prefix",
-        "300.0.0.0/8",       # Invalid IPv4
-        "2001:xyz::/32"      # Invalid IPv6
-    ]
-    for prefix in invalid_prefixes:
-        network = format_pfx(prefix)
-        if network:
-            assert is_bogon_pfx(network) is False
-        else:
-            assert network is None
